Remove old index.html
