+++ requirements.txt
@@ -4,3 +4,4 @@ pydantic
 python-multipart
 python-dotenv
 openai
+flask
