+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import os
+import sys
+
+# Ensure the project root is on the Python path so tests can import modules
+ROOT_DIR = os.path.dirname(os.path.dirname(__file__))
+if ROOT_DIR not in sys.path:
+    sys.path.insert(0, ROOT_DIR)
