+++ tests/test_cors.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from app import app
+from api import app
 
 
 client = TestClient(app)
